Allow passing URI arg to Async::HTTP::Internet
I recently performed memory profiling on an app built with 'async'.
I discovered URI object "manipulation" was using ~50% app memory, mostly
because of the mistakes in my code.
I'm more conscious of the way I handle URI's since.

This PR is a performance optimization for Endpoint.parse. It avoids
parsing a URI if URI object is already passed.

To enable this optimization we allow passing URI objects to
Async::HTTP::Internet.